utils/auth.py
import jwt
import os
from datetime import datetime, timedelta

SECRET = os.getenv("JWT_SECRET", "your_jwt_secret")

def generate_token(user_id):
    payload = {
        "user_id": user_id,
        "exp": datetime.utcnow() + timedelta(days=7)
    }
    return jwt.encode(payload, SECRET, algorithm="HS256")

def verify_token(token):
    try:
        return jwt.decode(token, SECRET, algorithms=["HS256"])
    except:
        return None