requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==2.3.1
python-dotenv==1.0.0
tensorflow==2.16.1
Pillow==10.2.0
requests==2.31.0
PyJWT==2.10.1